pyproject.toml
[project]
name = "team5-waterandfish-be"
version = "0.1.0"
description = ""
authors = [
    {name = "sweeetpotaooo",email = "sehyun5004@naver.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "motor (>=3.3.0,<4.0.0)",
    "pymongo (>=4.6.0,<5.0.0)",
    "PyJWT (>=2.8.0,<3.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "alembic (>=1.13.0,<2.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "team5_waterandfish_be", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"